feat(monkeys): Add support for running the monkeys in separate processes [WPB-6186]
