+++ monkeys/docker/Dockerfile
@@ -1,11 +1,12 @@
 FROM eclipse-temurin:17 AS jdk-build
 
 COPY .. /kalium
-RUN cd /kalium && ./gradlew :monkeys:build
+RUN cd /kalium && ./gradlew :monkeys:assembleDist
+RUN tar -xf /kalium/monkeys/build/distributions/monkeys.tar
 
 FROM wirebot/cryptobox:1.4.0
 
 RUN mkdir /opt/app
-COPY --from=jdk-build /kalium/monkeys/build/libs/monkeys.jar /opt/app
-ENTRYPOINT ["java", "-jar", "/opt/app/monkeys.jar"]
+COPY --from=jdk-build /monkeys /opt/app
+ENTRYPOINT ["/opt/app/bin/monkeys"]
 CMD ["/config.json"]
